+++ migrations/versions/f6fc95ece26a_.py
@@ -1,0 +1,24 @@
+"""empty message
+
+Revision ID: f6fc95ece26a
+Revises: 197ae256228d, 15218a083889
+Create Date: 2020-04-14 20:52:21.210139
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'f6fc95ece26a'
+down_revision = ('197ae256228d', '15218a083889')
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass