+++ src/pages/Submissions.vue
@@ -32,6 +32,12 @@
                     <icon name="gear"/>
                 </b-button>
 
+                <b-button v-if="canEmailStudents"
+                          v-b-popover.top.hover="`Email the authors of the visible submissions`"
+                          v-b-modal.submissions-page-email-students-modal>
+                    <icon name="envelope"/>
+                </b-button>
+
                 <b-button-group v-b-popover.bottom.hover="'Reload submissions'">
                     <submit-button :wait-at-least="500"
                                    name="refresh-button"
@@ -46,6 +52,39 @@
         <cg-logo v-if="isStudent"
                  :inverted="!darkMode" />
     </local-header>
+
+    <b-modal v-if="canEmailStudents"
+             id="submissions-page-email-students-modal"
+             ref="contactStudentModal"
+             size="xl"
+             hide-footer
+             no-close-on-backdrop
+             no-close-on-esc
+             hide-header-close
+             title="Email authors"
+             body-class="p-0"
+             dialog-class="auto-test-result-modal">
+        <cg-catch-error capture>
+            <template slot-scope="{ error }">
+                <b-alert v-if="error"
+                         show
+                         variant="danger">
+                    {{ $utils.getErrorMessage(error) }}
+                </b-alert>
+
+                <student-contact
+                    v-else
+                    :initial-users="visibleStudents"
+                    :course="assignment.course"
+                    :default-subject="defaultEmailSubject"
+                    reset-on-email
+                    @hide="() => $refs.contactStudentModal.hide()"
+                    @emailed="() => $refs.contactStudentModal.hide()"
+                    :can-use-snippets="canUseSnippets"
+                    class="p-3"/>
+            </template>
+        </cg-catch-error>
+    </b-modal>
 
     <div class="cat-container d-flex flex-column">
         <div v-if="error != null">
@@ -275,7 +314,7 @@
 
             <div v-if="selectedCat === 'analytics'"
                  class="flex-grow-1">
-                <catch-error>
+                <cg-catch-error>
                     <template slot-scope="scope">
                         <b-alert show variant="danger" v-if="scope.error">
                             An unexpected error occurred:
@@ -287,7 +326,7 @@
                         <analytics-dashboard v-else
                                              :assignment-id="assignmentId" />
                     </template>
-                </catch-error>
+                </cg-catch-error>
             </div>
         </template>
     </div>
@@ -308,6 +347,7 @@ import 'vue-awesome/icons/users';
 import 'vue-awesome/icons/chevron-down';
 import 'vue-awesome/icons/code-fork';
 import 'vue-awesome/icons/git';
+import 'vue-awesome/icons/envelope';
 
 import ltiProviders from '@/lti_providers';
 import { NONEXISTENT } from '@/constants';
@@ -317,7 +357,6 @@ import {
     CgLogo,
     Loader,
     Collapse,
-    CatchError,
     LocalHeader,
     CGIgnoreFile,
     RubricEditor,
@@ -329,6 +368,7 @@ import {
     SubmissionsExporter,
     WebhookInstructions,
 } from '@/components';
+import StudentContact from '@/components/StudentContact';
 
 import { setPageTitle, pageTitleSep } from './title';
 
@@ -349,7 +389,10 @@ export default {
     },
 
     computed: {
-        ...mapGetters('user', { userId: 'id' }),
+        ...mapGetters('user', {
+            userId: 'id',
+            userPerms: 'permissions',
+        }),
         ...mapGetters('pref', ['darkMode']),
         ...mapGetters('courses', ['assignments']),
         ...mapGetters('rubrics', { allRubrics: 'rubrics' }),
@@ -642,6 +685,36 @@ export default {
             // When there is a group, only show the add button if you are not a student.
             return !this.isStudent;
         },
+
+        visibleStudents() {
+            const seen = new Set();
+            return this.filteredSubmissions.reduce((acc, s) => {
+                s.user.getContainedUsers().forEach(u => {
+                    if (seen.has(u.id)) {
+                        return;
+                    }
+
+                    seen.add(u.id);
+                    acc.push(u);
+                });
+
+                return acc;
+            }, []);
+        },
+
+        defaultEmailSubject() {
+            return `[CodeGrade - ${this.assignment.course.name}/${this.assignment.name}] …`;
+        },
+
+        canUseSnippets() {
+            return !!this.userPerms.can_use_snippets;
+        },
+
+        canEmailStudents() {
+            return (UserConfig.features.email_students &&
+                    this.$utils.getProps(this.coursePermissions, false, 'can_email_students'));
+        },
+
     },
 
     watch: {
@@ -786,7 +859,6 @@ export default {
         CgLogo,
         Loader,
         Collapse,
-        CatchError,
         LocalHeader,
         CGIgnoreFile,
         RubricEditor,
@@ -798,6 +870,7 @@ export default {
         SubmissionsExporter,
         WebhookInstructions,
         LateSubmissionIcon,
+        StudentContact,
         AnalyticsDashboard: () => ({
             component: import('@/components/AnalyticsDashboard'),
             loading: Loader,